deployment/requirements.txt
aiofiles
certifi
click
cycler
dill
fastapi
h11
Jinja2
kiwisolver
loguru
MarkupSafe
matplotlib
numpy
opencv-python-headless
pandas
Pillow
pydantic
pyparsing
python-dateutil
python-multipart
pytz
six
starlette
torch-snippets
tqdm
typing-extensions
uvicorn
torch